+++ src/cpu/rv64/decode/inst_32/m_extension.rs
@@ -1,8 +1,8 @@
 use crate::cpu::instruction::OpecodeKind;
-use crate::cpu::rv32::decode::DecodeUtil;
+use crate::cpu::rv64::decode::DecodeUtil;
 use crate::cpu::TrapCause;
 
-pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
+pub fn parse_opecode(inst: u64) -> Result<OpecodeKind, &'static str> {
     let opmap: u8 = inst.slice(6, 0) as u8;
     let funct3: u8 = inst.slice(14, 12) as u8;
 
@@ -23,9 +23,9 @@ pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
 }
 
 pub fn parse_rd(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rd: usize = inst.slice(11, 7) as usize;
 
     match opkind {
@@ -42,9 +42,9 @@ pub fn parse_rd(
 }
 
 pub fn parse_rs1(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs1: usize = inst.slice(19, 15) as usize;
 
     match opkind {
@@ -61,9 +61,9 @@ pub fn parse_rs1(
 }
 
 pub fn parse_rs2(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs2: usize = inst.slice(24, 20) as usize;
 
     match opkind {
@@ -81,8 +81,8 @@ pub fn parse_rs2(
 
 #[allow(non_snake_case)]
 pub fn parse_imm(
-    _inst: u32,
+    _inst: u64,
     _opkind: &OpecodeKind,
-) -> Result<Option<i32>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<i32>, (Option<u64>, TrapCause, String)> {
     Ok(None)
 }
