+++ src/cpu/rv64/decode/inst_32/priv_extension.rs
@@ -1,8 +1,8 @@
 use crate::cpu::instruction::OpecodeKind;
-use crate::cpu::rv32::decode::DecodeUtil;
+use crate::cpu::rv64::decode::DecodeUtil;
 use crate::cpu::TrapCause;
 
-pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
+pub fn parse_opecode(inst: u64) -> Result<OpecodeKind, &'static str> {
     let _opmap: u8 = inst.slice(6, 0) as u8;
     let _funct3: u8 = inst.slice(14, 12) as u8;
     let funct7: u8 = inst.slice(31, 25) as u8;
@@ -19,16 +19,16 @@ pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
 }
 
 pub fn parse_rd(
-    _inst: u32,
+    _inst: u64,
     _opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     Ok(None)
 }
 
 pub fn parse_rs1(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs1: usize = inst.slice(19, 15) as usize;
 
     match opkind {
@@ -38,9 +38,9 @@ pub fn parse_rs1(
 }
 
 pub fn parse_rs2(
-    inst: u32,
+    inst: u64,
     opkind: &OpecodeKind,
-) -> Result<Option<usize>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs2: usize = inst.slice(24, 20) as usize;
 
     match opkind {
@@ -50,8 +50,8 @@ pub fn parse_rs2(
 }
 
 pub fn parse_imm(
-    _inst: u32,
+    _inst: u64,
     _opkind: &OpecodeKind,
-) -> Result<Option<i32>, (Option<u32>, TrapCause, String)> {
+) -> Result<Option<i32>, (Option<u64>, TrapCause, String)> {
     Ok(None)
 }
